Create new netty-graal module for native image configuration The graal configuration that netty needs was previously only present in the http-netty module (HttpNettyFeature). This patch creates a new netty-graal module for that configuration, so that other modules that use netty but not http-netty (in particular micronaut-grpc) can use the same configuration.
Resolves #6650
