+++ buffer-netty/build.gradle
@@ -6,6 +6,7 @@ internalSanityChecks {
 dependencies {
     api project(":core")
     api project(":inject")
+    api project(":netty-graal")
     api libs.managed.netty.buffer
 
     annotationProcessor project(":inject-java")
