+++ gradle/libs.versions.toml
@@ -277,6 +277,7 @@ managed-neo4j = { module = "org.neo4j.test:neo4j-harness", version.ref = "manage
 managed-neo4j-bolt = { module = "org.neo4j.driver:neo4j-java-driver", version.ref = "managed-neo4j-java-driver" }
 
 managed-netty-buffer = { module = "io.netty:netty-buffer", version.ref = "managed-netty" }
+managed-netty-common = { module = "io.netty:netty-common", version.ref = "managed-netty" }
 managed-netty-codec-http = { module = "io.netty:netty-codec-http", version.ref = "managed-netty" }
 managed-netty-codec-http2 = { module = "io.netty:netty-codec-http2", version.ref = "managed-netty" }
 managed-netty-handler = { module = "io.netty:netty-handler", version.ref = "managed-netty" }
