+++ settings.gradle
@@ -52,6 +52,7 @@ include "jackson-databind"
 include "json-core"
 include "management"
 include "messaging"
+include 'netty-graal'
 include "router"
 include "runtime"
 include "runtime-osx"
